# static/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The Theseus is an off-chain service which provides access to community entities (posts, likes, follows)",
    "title": "Theseus",
    "version": "1.2.1"
  },
  "basePath": "/v1",
  "paths": {
    "/ddv/stats": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "DDV"
        ],
        "summary": "Returns DDV stats.",
        "operationId": "GetDDVStats",
        "responses": {
          "200": {
            "description": "Stats",
            "schema": {
              "$ref": "#/definitions/DDVStats"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/posts": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Community"
        ],
        "summary": "Return posts with additional meta information.",
        "operationId": "ListPosts",
        "parameters": [
          {
            "maximum": 9,
            "minimum": 1,
            "example": 4,
            "description": "filters posts by category",
            "name": "category",
            "in": "query"
          },
          {
            "enum": [
              "created_at",
              "likesCount",
              "dislikesCount",
              "pdv"
            ],
            "type": "string",
            "default": "createdAt",
            "example": "likesCount",
            "description": "sets posts' field to be sorted by",
            "name": "sortBy",
            "in": "query"
          },
          {
            "enum": [
              "asc",
              "desc"
            ],
            "type": "string",
            "default": "desc",
            "example": "asc",
            "description": "sets sort's direct",
            "name": "orderBy",
            "in": "query"
          },
          {
            "example": "furya1ltx6yymrs8eq4nmnhzfzxj6tspjuymh8mgd6gz",
            "description": "filters posts by owner",
            "name": "owner",
            "in": "query"
          },
          {
            "example": "furya1ltx6yymrs8eq4nmnhzfzxj6tspjuymh8mgd6gz",
            "description": "filters posts by one who liked its",
            "name": "likedBy",
            "in": "query"
          },
          {
            "example": "furya1ltx6yymrs8eq4nmnhzfzxj6tspjuymh8mgd6gz",
            "description": "filters post by owners who followed by followedBy",
            "name": "followedBy",
            "in": "query"
          },
          {
            "maximum": 100,
            "minimum": 1,
            "default": 20,
            "description": "limits count of returned posts",
            "name": "limit",
            "in": "query"
          },
          {
            "example": "furya1ltx6yymrs8eq4nmnhzfzxj6tspjuymh8mgd6gz/df870e39-6fcb-11eb-9461-0242ac11000b",
            "description": "sets not-including bound for list by post id(`owner/uuid`)",
            "name": "after",
            "in": "query"
          },
          {
            "example": 1613414389,
            "description": "sets lower datetime bound for list",
            "name": "from",
            "in": "query"
          },
          {
            "example": 1613424389,
            "description": "sets upper datetime bound for list",
            "name": "to",
            "in": "query"
          },
          {
            "example": "furya1ltx6yymrs8eq4nmnhzfzxj6tspjuymh8mgd6gz",
            "description": "adds liked flag to response",
            "name": "requestedBy",
            "in": "query"
          },
          {
            "description": "excludes posts with negative pdv",
            "name": "excludeNegative",
            "in": "query"
          },
          {
            "description": "excludes posts with pdv = 0",
            "name": "excludeNeutral",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Posts",
            "schema": {
              "$ref": "#/definitions/ListPostsResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/posts/{owner}/{uuid}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Community"
        ],
        "summary": "Get post by owner and uuid.",
        "operationId": "GetPostByID",
        "parameters": [
          {
            "type": "string",
            "name": "owner",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "uuid",
            "in": "path",
            "required": true
          },
          {
            "example": "furya1ltx6yymrs8eq4nmnhzfzxj6tspjuymh8mgd6gz",
            "description": "adds liked flag to response",
            "name": "requestedBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Post",
            "schema": {
              "$ref": "#/definitions/GetPostResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "post not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/posts/{slug}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Community"
        ],
        "summary": "Get post by slug.",
        "operationId": "GetPostBySlug",
        "parameters": [
          {
            "type": "string",
            "name": "slug",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Post",
            "schema": {
              "$ref": "#/definitions/SharePost"
            }
          },
          "404": {
            "description": "post not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/profiles/stats": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Profiles"
        ],
        "summary": "Returns furya stats.",
        "operationId": "GetFuryaStats",
        "responses": {
          "200": {
            "description": "Stats",
            "schema": {
              "$ref": "#/definitions/FuryaStats"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/profiles/{address}/stats": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Profiles"
        ],
        "summary": "Get pdv stats by address.",
        "operationId": "GetProfileStats",
        "parameters": [
          {
            "type": "string",
            "name": "address",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Profile stats",
            "schema": {
              "$ref": "#/definitions/ProfileStats"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "profile not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Category": {
      "type": "integer",
      "format": "int32",
      "x-go-package": "github.com/TessorNetwork/furya/x/community/types"
    },
    "DDVStats": {
      "type": "object",
      "title": "DDVStats ...",
      "properties": {
        "stats": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StatsItem"
          },
          "x-go-name": "Stats"
        },
        "total": {
          "type": "number",
          "format": "double",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "FuryaStats": {
      "type": "object",
      "title": "FuryaStats ...",
      "properties": {
        "adv": {
          "type": "number",
          "format": "double",
          "x-go-name": "ADV"
        },
        "ddv": {
          "type": "number",
          "format": "double",
          "x-go-name": "DDV"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "Error": {
      "type": "object",
      "title": "Error ...",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "github.com/TessorNetwork/go-api"
    },
    "GetPostResponse": {
      "type": "object",
      "title": "GetPostResponse ...",
      "properties": {
        "post": {
          "$ref": "#/definitions/Post"
        },
        "profileStats": {
          "$ref": "#/definitions/ProfileStats"
        },
        "stats": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StatsItem"
          },
          "x-go-name": "Stats"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "LikeWeight": {
      "type": "integer",
      "format": "int32",
      "x-go-package": "github.com/TessorNetwork/furya/x/community/types"
    },
    "ListPostsResponse": {
      "type": "object",
      "title": "ListPostsResponse ...",
      "properties": {
        "posts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Post"
          },
          "x-go-name": "Posts"
        },
        "profileStats": {
          "description": "ProfileStats contains profiles stats.",
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/ProfileStats"
          },
          "x-go-name": "ProfileStats"
        },
        "stats": {
          "description": "Posts' statistics dictionary where key is a full form ID (owner/uuid) and value is statistics",
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/StatsItem"
            }
          },
          "x-go-name": "Stats"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "Post": {
      "type": "object",
      "title": "Post ...",
      "properties": {
        "category": {
          "$ref": "#/definitions/Category"
        },
        "createdAt": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "CreatedAt"
        },
        "dislikesCount": {
          "type": "integer",
          "format": "uint32",
          "x-go-name": "DislikesCount"
        },
        "likeWeight": {
          "$ref": "#/definitions/LikeWeight"
        },
        "likesCount": {
          "type": "integer",
          "format": "uint32",
          "x-go-name": "LikesCount"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "pdv": {
          "type": "number",
          "format": "double",
          "x-go-name": "PDV"
        },
        "previewImage": {
          "type": "string",
          "x-go-name": "PreviewImage"
        },
        "slug": {
          "type": "string",
          "x-go-name": "Slug"
        },
        "text": {
          "type": "string",
          "x-go-name": "Text"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "ProfileStats": {
      "type": "object",
      "title": "ProfileStats ...",
      "properties": {
        "postsCount": {
          "type": "integer",
          "format": "uint16",
          "x-go-name": "PostsCount"
        },
        "stats": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StatsItem"
          },
          "x-go-name": "Stats"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "SharePost": {
      "type": "object",
      "title": "SharePost ...",
      "properties": {
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    },
    "StatsItem": {
      "description": "Key is RFC3999 date, value is PDV.",
      "type": "object",
      "title": "StatsItem ...",
      "properties": {
        "date": {
          "type": "string",
          "x-go-name": "Date"
        },
        "value": {
          "type": "number",
          "format": "double",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/TessorNetwork/theseus/internal/server"
    }
  }
}